+++ .circleci/config.yml
@@ -147,7 +147,7 @@ jobs:
           name: Create release build of libwasmvm
           command: make build-rust
       - persist_to_workspace:
-          root: ./api
+          root: ./internal/api
           paths:
             - libwasmvm.x86_64.so
       - save_cache:
@@ -173,7 +173,7 @@ jobs:
           at: /tmp/builds
       - run:
           name: Copy .so build
-          command: cp /tmp/builds/libwasmvm.x86_64.so ./api
+          command: cp /tmp/builds/libwasmvm.x86_64.so ./internal/api
       - run:
           name: Go integration tests
           command: make test
@@ -190,7 +190,7 @@ jobs:
       - run: make test-alpine
       - run:
           name: Debug build results
-          command: ls -l ./api
+          command: ls -l ./internal/api
 
   deploy_to_git:
     machine:
@@ -211,7 +211,7 @@ jobs:
           command: make release-build-macos
       - run:
           name: Debug build results
-          command: ls -l ./api
+          command: ls -l ./internal/api
       - run:
           name: Configure git user
           # This is not a GitHub user and no permissions can be configured other than "push access", which
@@ -223,7 +223,7 @@ jobs:
           name: Check-in and push new libraries
           command: |
             git status
-            git add ./api
+            git add ./internal/api
             git commit --allow-empty -m '[skip ci] Built release libraries'
             git push origin $CIRCLE_BRANCH
 
@@ -238,13 +238,13 @@ jobs:
           command: make release-build-alpine
       - run:
           name: Debug build results
-          command: ls -l ./api
+          command: ls -l ./internal/api
       - run:
           name: Collect artifacts
           command: |
             mkdir artifacts
-            cp ./api/libwasmvm_muslc.a artifacts/libwasmvm_muslc.x86_64.a
-            cp ./api/libwasmvm_muslc.aarch64.a artifacts/libwasmvm_muslc.aarch64.a
+            cp ./internal/api/libwasmvm_muslc.a artifacts/libwasmvm_muslc.x86_64.a
+            cp ./internal/api/libwasmvm_muslc.aarch64.a artifacts/libwasmvm_muslc.aarch64.a
       - run:
           name: Create checksums
           working_directory: artifacts
