+++ internal/api/memory.go
