+++ internal/api/testdata/README.md
@@ -1,7 +1,7 @@
 Update contracts via e.g.
 
 ```sh
-cd api/testdata
+cd internal/api/testdata
 ./download_releases.sh v0.14.0-beta2
 ```
