+++ internal/api/testdata/download_releases.sh
